Fix bug where uncompressing key removes prefix
This method is supposed to remove the prefix if the argument
is false. However, it does so indiscriminately. As a result, it can
sometimes give the public key have 63 bytes. This commit fixes
it by checking to see if removing the prefix would result in the
public key having 63 bytes.